Rename: Update folder names to _components and _styles in 'explore, search, pointshop'
