+++ app/explore/_components/RecordSwitch.tsx
