+++ app/pointshop/_components/Cart.tsx
@@ -1,4 +1,4 @@
-import style from './pointShop.module.scss'
+import style from '../_styles/pointShop.module.scss'
 import PointIcon from './PointIcon';
 
 type Item = {
