+++ app/pointshop/_components/PointIcon.tsx
