+++ app/pointshop/_components/PointShop.tsx
@@ -1,5 +1,5 @@
 
-import style from '@/app/pointshop/_component/pointShop.module.scss'
+import style from '../_styles/pointShop.module.scss'
 import ShopInfoButton from './ShopInfoButton'
 import CustomProfile from './CustomProfile'
