+++ app/pointshop/_styles/shopInfo.module.scss
