+++ app/pointshop/i/flow/shopinfo/page.tsx
@@ -1,4 +1,4 @@
-import ShopInfo from "@/app/pointshop/_component/ShopInfo";
+import ShopInfo from "@/app/pointshop/_components/ShopInfo";
 
 export default function Page() {
     return (
