+++ app/pointshop/page.tsx
@@ -1,4 +1,4 @@
-import PointShop from '@/app/pointshop/_component/PointShop'
+import PointShop from '@/app/pointshop/_components/PointShop'
 
 export default function Page() {
     return (
