+++ app/search/_components/CategoryDropdown.tsx
@@ -1,18 +1,18 @@
 'use client'
-import style from './dropdown.module.scss'
+import style from '../_styles/dropdown.module.scss'
 
 type Props = {
-    sort: string,
-    setSortHandler: (item: string) => void, // 문자열을 받아야 합니다.
+    currentCategory: string,
+    setCurrentCategoryHandler: (item: string) => void, // 문자열을 받아야 합니다.
 }
 
-export default function SortDropdown({ sort, setSortHandler }: Props) {
-    const category = ['최신순', '추천순'];
+export default function CategoryDropdown({ currentCategory, setCurrentCategoryHandler }: Props) {
+    const category = ['전체', '썸', '연애', '짝사랑', '이별', '19'];
 
     return (
         <ul className={style.items}>
             {category.map((item, idx) => {
-                const isSelected = item === sort; // 현재 항목이 선택된 정렬과 일치하는지 확인
+                const isSelected = item === currentCategory; // 현재 항목이 선택된 카테고리와 일치하는지 확인
 
                 return (
                     <li
@@ -25,7 +25,7 @@ export default function SortDropdown({ sort, setSortHandler }: Props) {
                                     ? `${style.item} ${style.firstItem}`
                                     : `${style.item}`
                         }
-                        onClick={() => setSortHandler(item)}
+                        onClick={() => setCurrentCategoryHandler(item)}
                     >
                         {item}
                     </li>
