+++ app/search/_styles/dropdown.module.scss
