+++ app/search/_styles/search.module.scss
