+++ app/search/_styles/searchBar.module.scss
