+++ app/search/_styles/searchResult.module.scss
