--- app/search/layout.tsx
@@ -1,15 +1,0 @@
-import { ReactNode } from "react"
-
-type Props = { children: ReactNode };
-
-export default function Layout({ children }: Props) {
-
-    return (
-        <div>
-            {children}
-        </div>
-
-
-    )
-}
-
